rg-253: * move header to ResumesHeader
