+++ frontend/src/bundles/common/components/header-wrapper/components/resumes-header/styles.module.scss
@@ -1,0 +1,5 @@
+.resume_header__container {
+    display: flex;
+    align-items: center;
+    gap: 24px;
+}
