+++ frontend/src/bundles/common/components/header-wrapper/constants/constants.ts
@@ -1,2 +1,3 @@
 export { GetStartedHeaderRoutes } from './get-started-header-routes';
 export { NoHeaderRoutes } from './no-header-routes';
+export { ResumesHeaderRoutes } from './resumes-header-routes';
