+++ frontend/src/bundles/common/components/header-wrapper/constants/resumes-header-routes.ts
@@ -1,0 +1,5 @@
+import { AppRoute } from '~/bundles/common/enums/app-route.enum';
+
+const ResumesHeaderRoutes = [AppRoute.RESUME];
+
+export { ResumesHeaderRoutes };
