+++ frontend/src/bundles/resume/pages/styles.module.scss
@@ -10,12 +10,6 @@
     padding: 25px 100px;
 }
 
-.resume__actions {
-    display: flex;
-    align-items: center;
-    gap: 24px;
-}
-
 .share__icon {
     width: 25px;
     height: 25px;
